🚸 fix dashboard loading ui #28
