+++ src/app/(dashboard)/dashboard/skeletons.tsx
@@ -1,0 +1,5 @@
+import { Skeleton } from "@mantine/core";
+
+export function UserInfoSkeleton() {
+  return <Skeleton w={{ base: "auto" }} h={{ base: 324, md: 395 }} />;
+}
